requirements.txt
Flask
psycopg2
python-raven
PyYAML
